+++ components/primary.js
@@ -1,17 +1,17 @@
 import Image from "next/image"
 
-import { IDE } from "../path"
+import { PRIMARY } from "../constants/path"
 
-export function Ide () {
+export function Primary () {
     return (
         <div className=" inline-flex justify-start items-center">
             <h4 
                 className="text-md px-6 mr-10 w-1/4 border-r-2 text-gray-800 dark:text-gray-200 max-w-xl md:text-xl"
             >
-                I.D.E.
+                Primary
             </h4>
             <div className="flex flex-wrap gap-3 w-3/4">
-                { IDE.map(
+                { PRIMARY.map(
                     (TECH_STACKS) => ( 
                         <div
                             key={ TECH_STACKS.id }
