+++ styles/globals.css
@@ -41,5 +41,5 @@ a {
 }
 @font-face {
     font-family: "burtons";
-    src: url("../public/Burtons.otf");
+    src: url("/Burtons.otf");
 }
